feat: lab 1.2
done

+++ prj/WebDev/src/main/java/com/webdev/sdc/service/CurrencyService.java
@@ -1,5 +1,9 @@
 package com.webdev.sdc.service;
 
+import com.webdev.sdc.dto.CurrencyDto;
+import com.webdev.sdc.exception.ConflictException;
+import com.webdev.sdc.exception.NotFoundException;
+import com.webdev.sdc.exception.RandomException;
 import com.webdev.sdc.model.CurrencyEntity;
 import com.webdev.sdc.repository.CurrencyRepository;
 import org.springframework.stereotype.Service;
@@ -10,6 +14,7 @@ import java.util.Optional;
 @Service
 public class CurrencyService {
     private final CurrencyRepository repository;
+    private Long id = 1L;
 
     public CurrencyService(CurrencyRepository repository) {
         this.repository = repository;
@@ -27,11 +32,15 @@ public class CurrencyService {
         return repository.findById(id);
     }
 
-    public CurrencyEntity createCurrency(CurrencyEntity currency) {
-        return repository.save(currency);
+    public CurrencyEntity createCurrency(CurrencyDto currency) {
+        if (repository.findByType(currency.getType()).isEmpty()) {
+            return repository.save(this.entityFromDto(currency));
+        } else {
+            throw new ConflictException("Currency exists");
+        }
     }
 
-    public Optional<CurrencyEntity> updateCurrency(Long id, CurrencyEntity currency) {
+    public Optional<CurrencyEntity> updateCurrency(Long id, CurrencyDto currency) {
         return repository.findById(id).map(existing -> {
             CurrencyEntity updated = new CurrencyEntity(id, currency.getType(), currency.getRate());
             return repository.save(updated);
@@ -39,10 +48,21 @@ public class CurrencyService {
     }
 
     public boolean deleteCurrency(Long id) {
+        if (Math.random() < 0.5) {
+            throw new RandomException();
+        }
         if (repository.findById(id).isPresent()) {
             repository.deleteById(id);
             return true;
         }
-        return false;
+        throw new NotFoundException(id, "Currency");
+    }
+
+    private CurrencyEntity entityFromDto(CurrencyDto dto) {
+        return new CurrencyEntity(this.getIdAndIncrement(), dto.getType(), dto.getRate());
+    }
+
+    private Long getIdAndIncrement() {
+        return this.id++;
     }
 }
